Check the code with ruff and mypy in CI
pyproject.toml configures ruff (pyflakes, pycodestyle, bugbear, isort,
pyupgrade and ruff's own rules, 130 columns in src) and mypy on the
package; a CI job runs both at pinned versions.

What they found is fixed rather than silenced: the report and RCB
decoders read their fields through generic, checked accessors instead
of a dozen type: ignore comments; ReportEntry.value says it may be a
DataAccessError, and open61850-mms subscribe now prints one instead of
failing on it; an unused channel in three_phase, a variable reused with
two types in the association decoder, SCL attributes that may be absent.

+++ src/open61850/comtrade.py
@@ -276,7 +276,7 @@ def resample(times: list[float], values: list[float], rate: int) -> list[float]:
     clean = [0.0 if math.isnan(v) else v for v in values]
     t0, span = times[0], times[-1] - times[0]
     out = []
-    for k in range(int(math.floor(span * rate + 1e-9)) + 1):
+    for k in range(math.floor(span * rate + 1e-9) + 1):
         t = t0 + k / rate
         j = bisect.bisect_right(times, t) - 1
         if j >= len(times) - 1:
